+++ client/src/components/utils/index.js
@@ -1,4 +1,4 @@
-import { NoEdgeNetworkTile } from "./NoEdgeNetworkTile";
+export { templates } from "./composed-templates";
 export { DownloadCopyButtonSet } from "./downloadCopyButtons/DownloadCopyButtonSet";
 export { downloadContent } from "./downloadCopyButtons/DownloadConfig";
 export { NoSecretsManagerTile } from "./NoSecretsManagerTile";